% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_processing.R
\name{compare_samples}
\alias{compare_samples}
\title{Compare sample IDs between OUTRIDER and FRASER datasets (UI version)}
\usage{
compare_samples(ods, fds)
}
\arguments{
\item{ods}{OUTRIDER dataset object}

\item{fds}{FRASER dataset object}
}
\value{
A tagList containing summary and mismatches (for use in renderUI)
}
\description{
Compare sample IDs between OUTRIDER and FRASER datasets (UI version)
}
