% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_processing.R
\name{filtered_expression_genes}
\alias{filtered_expression_genes}
\title{Extract genes expressed in OUTRIDER dataset}
\usage{
filtered_expression_genes(ods)
}
\arguments{
\item{ods}{The OUTRIDER dataset}
}
\value{
An object containing the genes expressed in in the dataset
}
\description{
Extract genes expressed in OUTRIDER dataset
}
