% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_processing.R
\name{generate_results}
\alias{generate_results}
\title{Generate OUTRIDER and FRASER results with optional padjust filtering}
\usage{
generate_results(ods, fds, padj_threshold = 0.05, merged = TRUE)
}
\arguments{
\item{ods}{An OUTRIDER dataset}

\item{fds}{A FRASER dataset}

\item{padj_threshold}{Optional adjusted p-value threshold for filtering (default = 0.05)}

\item{merged}{select to return a dataframe containing the samples and genes in both FRASER and OUTRIDER}
}
\value{
list containing OUTRIDER and FRASER result tables
}
\description{
Generate OUTRIDER and FRASER results with optional padjust filtering
}
