% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_panelApp.R
\name{get_genes_for_panel}
\alias{get_genes_for_panel}
\title{Fetch gene symbols from a specific gene panel}
\usage{
get_genes_for_panel(panel_id, region = c("uk", "aus"))
}
\arguments{
\item{panel_id}{The panel ID (numeric or character) to retrieve genes from.}

\item{region}{A character string: "uk" or "aus" (default is "uk").}
}
\value{
A character vector of gene symbols in the specified panel.
}
\description{
Fetch gene symbols from a specific gene panel
}
\examples{
panels <- fetch_all_panels("uk")
get_genes_for_panel(panel_id = panels$id[1], region = "uk")
}
