% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_processing.R
\name{update_bam_paths}
\alias{update_bam_paths}
\title{Update BAM file paths for OUTRIDER or FRASER objects}
\usage{
update_bam_paths(object, new_base_path)
}
\arguments{
\item{object}{An OUTRIDER or FRASER dataset object (ods or fds)}

\item{new_base_path}{The new directory where BAM files are located}
}
\value{
The updated dataset object with corrected BAM file paths
}
\description{
Update BAM file paths for OUTRIDER or FRASER objects
}
