% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_processing.R
\name{generate_results}
\alias{generate_results}
\title{Generate OUTRIDER and FRASER results with optional padjust filtering}
\usage{
generate_results(
  ods,
  fds,
  padj_out = 0.05,
  padj_fra = 0.05,
  merged = TRUE,
  all = FALSE
)
}
\arguments{
\item{ods}{An OUTRIDER dataset (ODS)}

\item{fds}{A FRASER dataset (FDS)}

\item{padj_out}{Adjusted p-value threshold for OUTRIDER results (default = 0.05)}

\item{padj_fra}{Adjusted p-value threshold for FRASER results (default = 0.05)}

\item{merged}{Logical; whether to return a merged data.frame containing overlapping samples and genes (default = TRUE)}

\item{all}{Logical; whether to also return unfiltered result tables from OUTRIDER and FRASER (default = FALSE)}
}
\value{
A list containing filtered `outres` and `frares`, and optionally `merged`, `outres_all`, `frares_all`
}
\description{
Generate OUTRIDER and FRASER results with optional padjust filtering
}
