% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_gene_expression_multi}
\alias{plot_gene_expression_multi}
\title{Generate a faceted plot of gene expression for multiple genes}
\usage{
plot_gene_expression_multi(ods, genes, sampleID)
}
\arguments{
\item{ods}{An OUTRIDER dataset}

\item{genes}{A character vector of gene symbols or IDs}

\item{sampleID}{A sample ID to highlight in red}
}
\value{
A ggplot object with facets for each gene, highlighting the sample
}
\description{
Generate a faceted plot of gene expression for multiple genes
}
