% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_processing.R
\name{summarise_data}
\alias{summarise_data}
\title{Generate a table summarising available data and samples}
\usage{
summarise_data(ods, fds)
}
\arguments{
\item{ods}{An OUTRIDER dataset}

\item{fds}{A FRASER dataset}
}
\value{
data.frame with which data is available for each sample
}
\description{
Generate a table summarising available data and samples
}
